requirements.txt
langchain>=0.2.16
streamlit>=1.38.0
pandas>=2.2.2
psycopg2-binary